+++ lib/posix/posix_macos_amd64.nim
@@ -572,7 +572,7 @@ when defined(macosx):
     MSG_NOSIGNAL* = 0'i32
   var
     SO_NOSIGPIPE* {.importc, header: "<sys/socket.h>".}: cint
-elif defined(solaris):
+elif defined(sunos):
   # Solaris doesn't have MSG_NOSIGNAL
   const
     MSG_NOSIGNAL* = 0'i32